requirements.txt
nbox[serving]
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.1+cu113
Pillow
torchvision==0.13.1+cu113
git+https://github.com/huggingface/diffusers.git
accelerate==0.12.0
OmegaConf
wget
pytorch_lightning
huggingface_hub
ftfy
transformers
pyfiglet
triton==2.0.0.dev20220701
bitsandbytes
python-slugify
requests
tensorboard
https://github.com/apolinario/xformers/releases/download/0.0.2/xformers-0.0.14.dev0-cp38-cp38-linux_x86_64.whl